Commit integrador

+++ [UTN.Programacion]-Java/[Java]-Leccion1/src/Ejercicio3/Ejercicio3.java
@@ -4,17 +4,17 @@ import java.util.Scanner;
 
 public class Ejercicio3 {
     public static void main(String[] args) {
-	
-		//=========================CLASE 8 - VIERNES-27/05/2022=========================*/
+		/**============CLASE 8 - VIERNES-27/05/2022============*/
+        //Sacar área y perímetro de un rectángulo
         //Fórmula: Área = alto*ancho;
         //Fórmula: Perímetro=(alto+ancho)*2;
 
         Scanner leer = new Scanner(System.in);
 
         System.out.print("Ingrese alto: ");
-        int alto= Integer.parseInt(leer.nextLine());
+        float alto= Float.parseFloat(leer.nextLine());
         System.out.print("Ingrese ancho: ");
-        int ancho= Integer.parseInt(leer.nextLine());
+        float ancho= Float.parseFloat(leer.nextLine());
         System.out.println("El área es: "+(alto*ancho));
         System.out.println("El perímetro es: "+((alto+ancho)*2));
